+++ models/internal.go
@@ -25,14 +25,15 @@ type FrigateConnection struct {
 }
 
 type Notifiers struct {
-	Enabled  bool             `json:"enabled" example:"true" doc:"State of whether Frigate-Notify is enabled for notifications"`
-	Discord  []NotifierStatus `json:"discord" doc:"Status of Discord notifications"`
-	Gotify   []NotifierStatus `json:"gotify" doc:"Status of Gotify notifications"`
-	Ntfy     []NotifierStatus `json:"ntfy" doc:"Status of Ntfy notifications"`
-	Pushover []NotifierStatus `json:"pushover" doc:"Status of Pushover notifications"`
-	SMTP     []NotifierStatus `json:"smtp" doc:"Status of SMTP notifications"`
-	Telegram []NotifierStatus `json:"telegram" doc:"Status of Telegram notifications"`
-	Webhook  []NotifierStatus `json:"webhook" doc:"Status of Webhook notifications"`
+	Enabled    bool             `json:"enabled" example:"true" doc:"State of whether Frigate-Notify is enabled for notifications"`
+	Discord    []NotifierStatus `json:"discord" doc:"Status of Discord notifications"`
+	Gotify     []NotifierStatus `json:"gotify" doc:"Status of Gotify notifications"`
+	Ntfy       []NotifierStatus `json:"ntfy" doc:"Status of Ntfy notifications"`
+	Pushover   []NotifierStatus `json:"pushover" doc:"Status of Pushover notifications"`
+	SMTP       []NotifierStatus `json:"smtp" doc:"Status of SMTP notifications"`
+	Telegram   []NotifierStatus `json:"telegram" doc:"Status of Telegram notifications"`
+	Webhook    []NotifierStatus `json:"webhook" doc:"Status of Webhook notifications"`
+	Mattermost []NotifierStatus `json:"mattermost" doc:"Status of Mattermost notifications"`
 }
 
 type NotifierStatus struct {